完善模型
