+++ data/test_data/test.txt
@@ -1,4 +1,6 @@
-不含输液器具-***-不胶输液器具
-一曲长歌一悠回，-***-一曲长菖一悠回，
-雨打青桥折伞迎。-***-雨打青桥折伞迎啶
-风转度回伊媚城。-***-风转度回衅媚城。
+上海汇城路-***-上约汇城路
+一曲长歌一悠回-***-一曲层歌一悠回
+汐界蓝灵载客归-***-汐界蓝灵载客匠
+乐尾双虹两岸潵-***-乐尾双虹邱岸潵
+若期长思飞江踏-***-若期长思飞江叱
+不问樊花滴漏子-***-不问樊花萝漏子
