+++ data/train_data/train.txt
@@ -8,33 +8,33 @@
 全自动生化分析仪
 南京银行阳光广场支行
 南昌市丰利医疗器械有限公司
-七梦朱窗嫣花红，
+七梦朱窗嫣花红
 弦音绕洲梁上笼
-琴瑟温携谁人听？
-风华掠宇与君同。
-糖化渺境惹尘烟，
-霖住陌上东风怜。
-舞夕方至青州浅，
-晓梦烛干曾相见。
-千花沁红散暗菲，
-泷水长柳落絮微。
-一曲长歌一悠回，
-汐界蓝灵载客归。
-欢鲤戏水舞荷花，
-乐尾双虹两岸潵。
-若期长思飞江踏，
-鱼跃龙门触天霞。
-沐云修虹天色晴，
-雨打青桥折伞迎。
-橙深藏处系旧景，
-风转度回伊媚城。
-夏荷布塘十里浦，
-义供轻烟千波簇。
-树幽斜阳万回路，
-君临风华一王途。
-何欢陌巷沾帘屏，
-慕颜尘醉取妆容。
-雪落朱庭点梦珑，
-吻尽琉花落迹空。
-不问樊花滴漏子，
-落红媚眼犹缘情。
+琴瑟温携谁人听
+风华掠宇与君同
+糖化渺境惹尘烟
+霖住陌上东风怜
+舞夕方至青州浅
+晓梦烛干曾相见
+千花沁红散暗菲
+泷水长柳落絮微
+一曲长歌一悠回
+汐界蓝灵载客归
+欢鲤戏水舞荷花
+乐尾双虹两岸潵
+若期长思飞江踏
+鱼跃龙门触天霞
+沐云修虹天色晴
+雨打青桥折伞迎
+橙深藏处系旧景
+风转度回伊媚城
+夏荷布塘十里浦
+义供轻烟千波簇
+树幽斜阳万回路
+君临风华一王途
+何欢陌巷沾帘屏
+慕颜尘醉取妆容
+雪落朱庭点梦珑
+吻尽琉花落迹空
+不问樊花滴漏子
+落红媚眼犹缘情
