+++ smbert/layers/SM_Bert_mlm.py
@@ -33,6 +33,7 @@ class SMBertMlm(nn.Module):
         # 申明网络
         self.smbert_emd = SMBbertEmbeddings(vocab_size=self.vocab_size, max_len=self.max_len, hidden_size=self.hidden_size)
         self.bi_gru = BiGRU(self.hidden_size, self.hidden_size)
+        self.gru_dense = nn.Linear(self.hidden_size, self.hidden_size)
         self.sigmoid = nn.Sigmoid()
         self.transformer_blocks = nn.ModuleList(
             Transformer(
@@ -64,6 +65,7 @@ class SMBertMlm(nn.Module):
         # error detection
         # 这里要严格符合GRU模块的输入size，内部已将batch_first改为True
         bi_gru_x = self.bi_gru(embedding_x)
+        bi_gru_x = self.gru_dense(bi_gru_x)
         pi = self.sigmoid(bi_gru_x)
         embedding_i = pi * mask_embedding_x + (1 - pi) * embedding_x
